Mantis 35857 : add opposite_route_id as foreign key

+++ db/migrate/20150304084602_remove_hub_restrictions_from_referentials.rb
@@ -1,5 +1,7 @@
 class RemoveHubRestrictionsFromReferentials < ActiveRecord::Migration
   def change
-    remove_column :referentials, :hub_restrictions
+	 if column_exists? :referentials, :hub_restrictions
+      remove_column :referentials, :hub_restrictions
+	 end
   end
 end
