+++ db/migrate/20150304084646_remove_hub_restrictions_by_default_from_organisations.rb
@@ -1,5 +1,7 @@
 class RemoveHubRestrictionsByDefaultFromOrganisations < ActiveRecord::Migration
   def change
-    remove_column :organisations, :hub_restrictions_by_default
+	 if column_exists? :organisations, :hub_restrictions_by_default
+      remove_column :organisations, :hub_restrictions_by_default
+	 end
   end
 end
